+++ server/metrics.go
@@ -2,16 +2,36 @@ package server
 
 import (
 	"crypto/tls"
+	"fmt"
 	"net/http"
 	"strings"
 	"time"
 
+	"github.com/dgrijalva/jwt-go"
 	"github.com/gin-gonic/gin"
+
+	oauth "github.com/meateam/api-gateway/oauth"
 	"github.com/meateam/api-gateway/user"
 	es "github.com/olivere/elastic/v7"
 	"github.com/spf13/viper"
 	"go.elastic.co/apm"
 )
+
+const (
+	// NoAuthType states thate no auth type was declared in the headers
+	NoAuthType = "NoneAuthType"
+
+	// UnknownAppID is the app id given to unknown app ids
+	UnknownAppID = "UnknownAppID"
+
+	// ClientNameLabel is the claim name of the client name of the requesting external application
+	ClientNameLabel = "currentUnit"
+)
+
+type extractedTokenInfo struct {
+	appID    string
+	authType string
+}
 
 type body struct {
 	User      *user.User `json:"user,omitempty"`
@@ -20,6 +40,8 @@ type body struct {
 	TimeStamp time.Time  `json:"timestamp,omitempty"`
 	Date      time.Time  `json:"date,omitempty"`
 	TraceID   string     `json:"traceID,omitempty"`
+	AuthType  string     `json:"authType,omitempty"`
+	AppID     string     `json:"appID,omitempty"`
 }
 
 // NewMetricsLogger initializes the metrics middleware.
@@ -35,19 +57,25 @@ func NewMetricsLogger() gin.HandlerFunc {
 
 		currentTransaction := apm.TransactionFromContext(c.Request.Context())
 
+		reqInfo := extractRequestInfo(c)
+
 		t := time.Now()
 		roundedDate := time.Date(t.Year(), t.Month(), t.Day(), 0, 0, 0, 0, t.Location()) // round the time to date (day-month-year)
 
+		matricsBson := &body{
+			User:      user.ExtractRequestUser(c),
+			Path:      c.Request.URL.Path,
+			Method:    c.Request.Method,
+			TimeStamp: time.Now(),
+			Date:      roundedDate,
+			TraceID:   currentTransaction.TraceContext().Trace.String(),
+			AuthType:  reqInfo.authType,
+			AppID:     reqInfo.appID,
+		}
+
 		_, _ = client.Index().
 			Index(index).
-			BodyJson(&body{
-				User:      user.ExtractRequestUser(c),
-				Path:      c.Request.URL.Path,
-				Method:    c.Request.Method,
-				TimeStamp: time.Now(),
-				Date:      roundedDate,
-				TraceID:   currentTransaction.TraceContext().Trace.String(),
-			}).
+			BodyJson(matricsBson).
 			Do(c.Request.Context())
 	}
 }
@@ -74,4 +102,39 @@ func initESConfig() ([]es.ClientOptionFunc, string) {
 	}
 
 	return elasticOpts, viper.GetString(configElasticsearchIndex)
+}
+
+func extractRequestInfo(c *gin.Context) *extractedTokenInfo {
+	var claims jwt.MapClaims
+	authType := c.GetHeader(oauth.AuthTypeHeader)
+
+	if authType == "" {
+		authType = "NoneAuthType"
+	} else {
+		spikeToken, err := oauth.ExtractTokenFromHeader(c)
+		if err != nil {
+			fmt.Printf("metrics token extraction error: %v", err)
+		}
+
+		spikeTokenObject, err := jwt.Parse(spikeToken, nil)
+		// Check the error is not because of nil keyfunc
+		if !((err != nil) && (err.(*jwt.ValidationError).Errors == jwt.ValidationErrorUnverifiable)) {
+			fmt.Printf("metrics token parsing error: %v", err)
+		}
+		claims, _ = spikeTokenObject.Claims.(jwt.MapClaims) //the token claims should conform to MapClaims
+	}
+
+	appID := oauth.DriveAppID
+
+	switch authType {
+	case oauth.DropboxAuthTypeValue:
+		appID = oauth.DropboxAppID
+	case oauth.CTSAuthTypeValue:
+		appID = oauth.CTSAppID
+	case oauth.ServiceAuthCodeTypeValue:
+		appID, _ = claims["clientName"].(string) // get the appID from the claims
+	}
+	finalInfo := &extractedTokenInfo{appID: appID, authType: authType}
+
+	return finalInfo
 }